refactor(race): Make entity folder for race widgets
- Made parent entity named folder for race widgets
