+++ packages/client/src/widgets/race/ui/index.ts
@@ -1,0 +1,1 @@
+export { RaceList } from './list';
