+++ packages/client/src/widgets/race/ui/list/index.ts
